feat(server): modify starting server files and add auth related packages
